requirements.txt
torch
diffusers
pytorch-lightning
transformers
bitsandbytes
tensorboard
matplotlib
scipy
SimpleITK
PyYAML